Moves sections/ to app folder to represent the home page's sections
