+++ app/sections/ContactSection.tsx
@@ -1,6 +1,6 @@
 import { Section, Text } from "@radix-ui/themes";
-import AccentedHeading from "../AccentedHeading";
-import CTAButton from "../CTAButton";
+import AccentedHeading from "../components/AccentedHeading";
+import CTAButton from "../components/CTAButton";
 
 const ContactSection: React.FC = () => {
   return (
